move test to specific folder
